Bug fixed: 1. Today information not presented as expected in Home VC 2. Dashboard picture cannot present if the nutrition is less than 0.5 3. Cannot cancel tapped elements in confirm photo VC 4. Update home view when standard has changed 5. In progress of fixing camera vc loading problem.

+++ Food Diary App!/Model/CoreDataHandler.swift
@@ -89,7 +89,7 @@ struct CoreDataHandler
             {
                 for result in results as! [NSManagedObject]
                 {
-                    if let queryElement = result.value(forKey: queryFor) as? Int
+                    if let queryElement = result.value(forKey: queryFor) as? Double
                     {
                         // Store data in the corresponding array
                         if queryElement != 0
